Restructured taxes.
