+++ src/analysis/tax_transfer.py
@@ -18,7 +18,7 @@ from src.analysis.tax_transfer_funcs.social_insurance import soc_ins_contrib
 from src.analysis.tax_transfer_funcs.taxes import favorability_check
 from src.analysis.tax_transfer_funcs.taxes import kindergeld
 from src.analysis.tax_transfer_funcs.taxes import tax_sched
-from src.analysis.tax_transfer_funcs.zve import zve
+from src.analysis.tax_transfer_funcs.taxes.zve import zve
 from src.model_code.imports import get_params
