+++ src/analysis/tax_transfer_funcs/test_tax_transfers/test_kindergeld.py
@@ -2,9 +2,9 @@ import pandas as pd
 import pytest
 from pandas.testing import assert_series_equal
 
-from src.analysis.tax_transfer_funcs.taxes import kg_eligibility_hours
-from src.analysis.tax_transfer_funcs.taxes import kg_eligibility_wage
-from src.analysis.tax_transfer_funcs.taxes import kindergeld
+from src.analysis.tax_transfer_funcs.taxes.kindergeld import kg_eligibility_hours
+from src.analysis.tax_transfer_funcs.taxes.kindergeld import kg_eligibility_wage
+from src.analysis.tax_transfer_funcs.taxes.kindergeld import kindergeld
 from src.analysis.tax_transfer_funcs.test_tax_transfers.auxiliary_test_tax import (
     load_input,
 )
