+++ src/analysis/tax_transfer_funcs/test_tax_transfers/test_tax_sched.py
@@ -1,6 +1,9 @@
-import pytest
 import pandas as pd
+import pytest
 from pandas.testing import assert_frame_equal
+
+from src.analysis.tax_transfer_funcs.taxes.calc_taxes import tarif
+from src.analysis.tax_transfer_funcs.taxes.calc_taxes import tax_sched
 from src.analysis.tax_transfer_funcs.test_tax_transfers.auxiliary_test_tax import (
     load_input,
 )
@@ -10,7 +13,6 @@ from src.analysis.tax_transfer_funcs.test_tax_transfers.auxiliary_test_tax impor
 from src.analysis.tax_transfer_funcs.test_tax_transfers.auxiliary_test_tax import (
     load_tb,
 )
-from src.analysis.tax_transfer_funcs.taxes import tax_sched, tarif
 
 input_cols = [
     "pid",
