+++ src/analysis/tax_transfer_ubi.py
@@ -12,7 +12,7 @@ from src.analysis.tax_transfer_funcs.social_insurance import soc_ins_contrib
 from src.analysis.tax_transfer_funcs.taxes import favorability_check
 from src.analysis.tax_transfer_funcs.taxes import soli
 from src.analysis.tax_transfer_funcs.taxes import tax_sched
-from src.analysis.tax_transfer_funcs.zve import zve
+from src.analysis.tax_transfer_funcs.taxes.zve import zve
 from src.model_code.imports import aggr
 from src.model_code.imports import get_params
